done authen

+++ src/modules/restaurants/restaurants.controller.ts
@@ -7,6 +7,8 @@ import {
   Put,
   Delete,
   Query,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 
 import { Query as ExpressQuery } from 'express-serve-static-core';
@@ -14,50 +16,62 @@ import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurant.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
+import { AuthGuard } from '@nestjs/passport';
+import { CurrentUser } from '../auth/decorators/current-user.decorator';
+import { User } from '../auth/schemas/user.schema';
+import { RolesGuard } from '../auth/guards/roles.guard';
+import { Roles } from '../auth/decorators/roles.decorator';
 
 @Controller('restaurants')
 export class RestaurantsController {
   constructor(private restaurantsService: RestaurantsService) {}
 
   @Get()
-  async getAllRestaurants(
-    @Query()
-    query: ExpressQuery,
-  ): Promise<Restaurant[]> {
+  @UseGuards(AuthGuard())
+  async getAllRestaurants(@Query() query: ExpressQuery): Promise<Restaurant[]> {
     return this.restaurantsService.findAll(query);
   }
 
   @Post()
+  @UseGuards(AuthGuard(), RolesGuard)
+  @Roles('admin', 'user')
   async createRestaurant(
-    @Body()
-    restaurant: CreateRestaurantDto,
+    @Body() restaurant: CreateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
-    return this.restaurantsService.create(restaurant);
+    return this.restaurantsService.create(restaurant, user);
   }
 
   @Get(':id')
-  async getRestaurantById(
-    @Param('id')
-    id: string,
-  ): Promise<Restaurant> {
+  @UseGuards(AuthGuard())
+  async getRestaurantById(@Param('id') id: string): Promise<Restaurant> {
     return this.restaurantsService.findById(id);
   }
 
   @Put(':id')
+  @UseGuards(AuthGuard())
   async updateRestaurantById(
-    @Param('id')
-    id: string,
-    @Body()
-    restaurant: UpdateRestaurantDto,
+    @Param('id') id: string,
+    @Body() restaurant: UpdateRestaurantDto,
+    @CurrentUser() user: User,
   ): Promise<Restaurant> {
+    const res = await this.restaurantsService.findById(id);
+    if (res.owner.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You can not update this restaurant');
+    }
     return this.restaurantsService.updateById(id, restaurant);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard())
   async deleteRestaurantById(
-    @Param('id')
-    id: string,
+    @Param('id') id: string,
+    @CurrentUser() user: User,
   ): Promise<Boolean> {
+    const res = await this.restaurantsService.findById(id);
+    if (res.owner.toString() !== user._id.toString()) {
+      throw new ForbiddenException('You can not delete this restaurant');
+    }
     return this.restaurantsService.deleteById(id);
   }
 }
